+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/ContentHeader.jsx
@@ -1,17 +1,13 @@
 import React from 'react';
 import Button from 'terra-button';
 import Header from 'terra-clinical-header';
-import classNames from 'classnames/bind';
-import styles from './ContentHeader.module.scss';
-
-const cx = classNames.bind(styles);
 
 const TitleHeader = () => (
   <div>
     <Header
-      startContent={<div><Button text="Button" className={cx('header-button')} /></div>}
+      startContent={<Button text="Button" />}
       title="Default Header"
-      endContent={<div><Button text="Button" className={cx('header-button')} /></div>}
+      endContent={<Button text="Button" />}
     />
   </div>
 );