+++ packages/terra-clinical-item-collection/src/terra-dev-site/test/clinical-item-collection/ItemCollectionSelectable.test.jsx
@@ -17,8 +17,8 @@ class DefaultItemCollection extends React.Component {
       <div>
         <h3>This a POC Selectable ItemCollection Example.</h3>
         <p>
-This example it set up to be a single-selection example, however the Item Collection
-           gives implementors the flexibilty to create their own selection implementation.
+          This example it set up to be a single-selection example, however the Item Collection
+          gives implementors the flexibility to create their own selection implementation.
         </p>
         <p id="selected-key">
           Selected Item:
@@ -32,6 +32,7 @@ This example it set up to be a single-selection example, however the Item Collec
           hasComment
           hasEndAccessory
           breakpoint="medium"
+          ariaLabel="Item Collection"
         >
           <ItemCollection.Item
             id="item1"