+++ packages/terra-clinical-item-display/src/ItemDisplay.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames/bind';
+import classNames from 'classnames';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
 import styles from './ItemDisplay.module.scss';
 
-const cx = classNames.bind(styles);
+const cx = classNamesBind.bind(styles);
 
 const TextStyles = {
   PRIMARY: 'primary',
@@ -45,8 +47,7 @@ const propTypes = {
 
 const defaultProps = {
   text: '',
-  // TODO: textStyle should be set to 'primary' on the next major version bump: https://github.com/cerner/terra-clinical/issues/526
-  textStyle: undefined,
+  textStyle: 'primary',
   isTruncated: false,
   isDisabled: false,
   icon: undefined,
@@ -62,16 +63,23 @@ const ItemDisplay = ({
   iconAlignment,
   ...customProps
 }) => {
-  const componentClassNames = cx([
-    'item-display',
-    { 'is-disabled': isDisabled },
-    { [`icon-${iconAlignment}`]: icon },
+  const theme = React.useContext(ThemeContext);
+  const componentClassNames = classNames(
+    cx(
+      'item-display',
+      { 'is-disabled': isDisabled },
+      { [`icon-${iconAlignment}`]: icon },
+      theme.className,
+    ),
     customProps.className,
-  ]);
+  );
   const textClassNames = cx([
     'text',
     { 'is-truncated': isTruncated },
-    { 'strike-through': textStyle === TextStyles.STRIKETHROUGH, [`${textStyle}`]: textStyle },
+    { [`${textStyle}`]: textStyle === TextStyles.SECONDARY },
+    { [`${textStyle}`]: textStyle === TextStyles.ATTENTION },
+    { [`${textStyle}`]: textStyle === TextStyles.STRONG },
+    { 'strike-through': textStyle === TextStyles.STRIKETHROUGH },
   ]);
 
   let displayIcon;
@@ -80,7 +88,7 @@ const ItemDisplay = ({
   }
 
   return (
-    <div {...customProps} className={componentClassNames}>
+    <div {...customProps} className={componentClassNames} aria-disabled={isDisabled}>
       {displayIcon}
       <div data-terra-clinical-item-display-text className={textClassNames}>{text}</div>
     </div>